requirements.txt
BzrTools==2.6.0
Django==1.11.11
Markdown==2.6.11
Pillow==5.1.0
SecretStorage==2.0.0
Unipath==1.1
argparse==1.2.1
bzr==2.7.0dev1
certifi==2018.1.18
chardet==3.0.4
colorama==0.2.5
configobj==4.7.2
django-cors-headers==2.2.0
django-filter==1.1.0
django-secure==1.0.1
django-storages==1.6.6
djangorestframework==3.8.2
dropbox==8.7.1
duplicity==0.6.23
html5lib==0.999
httplib2==0.8
idna==2.6
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
mercurial==2.8.2
mysqlclient==1.3.12
oauth==1.0.1
paramiko==1.10.1
pycrypto==2.6.1
pygobject==3.12.0
pygpgme==0.3
python-apt==0.9.3.5ubuntu2
pytz==2018.4
requests==2.18.4
simplejson==3.3.1
six==1.5.2
urllib3==1.22
wadllib==1.3.2
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5